Add test for stubServlet deploymentContext
